refactor: move location to its own dir
